requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
google-api-python-client==2.108.0
selenium==4.15.2
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
networkx==3.2.1
matplotlib==3.8.2
flask==2.3.3
python-dotenv==1.0.0
pandas==2.1.3
scikit-learn==1.3.2
tqdm==4.66.1
colorlog==6.7.0
webdriver-manager==4.0.1
scholarly==1.7.11
newspaper3k==0.2.8
serpapi==0.1.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2